+++ VoxReader/Interfaces/IPalette.cs
@@ -3,5 +3,6 @@ namespace VoxReader.Interfaces
     public interface IPalette
     {
         Color[] Colors { get; }
+        string[] Notes { get; }
     }
 }