+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -->
 
+## [Unreleased]
+
+### Added
+
+- Added support for parsing palette notes
+
 ## [3.1.0] - 2023-01-03
 
 ### Added